Refactor Telegram bot service and update build configuration
- Replaced Telegram bot library with `dev.inmo:tgbotapi` for improved functionality and performance.
- Updated `build.gradle.kts` to reflect new dependencies and removed deprecated Telegram bot implementations.
- Enhanced `TelegramBotService` to streamline message handling and integrate with the new bot library.
- Adjusted user ID handling across various services to consistently use `Long` instead of `Int` for better compatibility.
- Improved error handling and logging throughout the bot service for enhanced user experience.

+++ core_service/src/main/kotlin/com/jobsalerts/core/repository/JobSearchRepository.kt
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository
 @Repository
 interface JobSearchRepository : MongoRepository<JobSearchOut, String> {
     fun findByUserId(userId: Long): List<JobSearchOut>
-    fun findByIdAndUserId(id: String, userId: Int): JobSearchOut?
-    fun deleteByIdAndUserId(id: String, userId: Int): Long
+    fun findByIdAndUserId(id: String, userId: Long): JobSearchOut?
+    fun deleteByIdAndUserId(id: String, userId: Long): Long
 }
