+++ core_service/src/main/kotlin/com/jobsalerts/core/service/commands/AlertCreationService.kt
@@ -94,7 +94,7 @@ class AlertCreationService(
         try {
             sendMessage(chatId, "🔍 Analyzing your job alert description...")
 
-            val parseResult = jobSearchParserService.parseUserInput(description, userId.toInt())
+            val parseResult = jobSearchParserService.parseUserInput(description, userId)
 
             if (parseResult.success && parseResult.jobSearchIn != null) {
                 // Successfully parsed, show for confirmation
