+++ core_service/src/main/kotlin/com/jobsalerts/core/service/commands/EditSearchService.kt
@@ -200,7 +200,7 @@ class EditSearchService(
         try {
             sendMessage(chatId, "🔍 Analyzing your updated job search criteria...")
             
-            val parseResult = jobSearchParserService.parseUserInput(description, userId.toInt())
+                            val parseResult = jobSearchParserService.parseUserInput(description, userId)
             
             if (parseResult.success && parseResult.jobSearchIn != null) {
                 // Successfully parsed, show for confirmation
